// web/src/shared/api/generated/model/questionImportPreviewResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Elmanhg.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { QuestionImportRowError } from './questionImportRowError';
import type { QuestionImportTypeCount } from './questionImportTypeCount';

export interface QuestionImportPreviewResult {
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  totalRows: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  validRows: number | string;
  types: QuestionImportTypeCount[];
  errors: QuestionImportRowError[];
}